+++ packages/typespec-go/src/tcgcadapter/clients.ts
@@ -426,6 +426,16 @@ export class clientAdapter {
         throw new AdapterError('UnsupportedTsp', `paging with re-injected parameters is not supported`, sdkMethod.__raw?.node ?? NoTarget);
       }
       method = new go.PageableMethod(methodName, goClient, sdkMethod.operation.path, sdkMethod.operation.verb, statusCodes, naming);
+      if (sdkMethod.pagingMetadata.nextLinkVerb && sdkMethod.pagingMetadata.nextLinkVerb !== 'GET') {
+        // we default to GET in the ctor for PageableMethod
+        switch (sdkMethod.pagingMetadata.nextLinkVerb) {
+          case 'POST':
+            method.nextLinkVerb = 'post';
+            break;
+          default:
+            sdkMethod.pagingMetadata.nextLinkVerb satisfies never;
+        }
+      }
       if (sdkMethod.pagingMetadata.nextLinkSegments) {
         method.nextLinkName = capitalize(sdkMethod.pagingMetadata.nextLinkSegments.map((segment) => {
           if (segment.kind === 'property') {