+++ packages/typespec-go/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Features Added
 
 * Emit `LICENSE.txt` file (defaults to MIT license with a Microsoft Copyright).
+* Added support for `nextLinkVerb` for pageable operations.
 
 ### Other Changes